+++ src/jsonpath.js
@@ -594,7 +594,9 @@ JSONPath.prototype._trace = function (
             ));
         }
     // simple case--directly follow property
-    } else if (!literalPriority && val && hasOwnProp.call(val, loc)) {
+    } else if (
+        !literalPriority && (val || val === '') && hasOwnProp.call(val, loc)
+    ) {
         addRet(
             this._trace(x, val[loc], push(path, loc), val, loc, callback,
                 hasArrExpr, true)