+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Fix: Require `json` as "own" property
 - Fix: wrap: false returning inconsistent data types (@CacheControl)
 - Fix: Ensure throwing with a bad result type
+- Fix: Allow empty string keys
 - Fix: Avoid erring when value before parent selector is falsey
 - Fix: If `resultType` is "all", if path resolves internally to a
     non-array (string), ensure it is converted to an array before